pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shelvr"
version = "0.1.0"
description = "Self-hosted ebook library server — part of Ink & Iron Apps"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "AGPL-3.0-only" }
authors = [{ name = "Ink & Iron Apps", email = "support@inknironapps.com" }]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: FastAPI",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19",
    "alembic>=1.13",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-jose[cryptography]>=3.3",
    "passlib[argon2]>=1.7",
    "structlog>=24.1",
    "httpx>=0.27",
    "ebooklib>=0.18",
    "pymupdf>=1.24",
    "mobi>=0.3.3",
    "Pillow>=10.3",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.9",
]

[project.scripts]
shelvr = "shelvr.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["shelvr*"]
exclude = ["tests*", "plugins*", "docs*", "alembic*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --strict-config"
markers = [
    "integration: integration tests that touch the DB or network",
]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "UP",
    "ASYNC",
    "C4",
    "SIM",
]
ignore = [
    "E501",
]

[tool.ruff.lint.per-file-ignores]
# FastAPI's Depends() is idiomatically used as a default argument value.
"shelvr/api/**/*.py" = ["B008"]
"shelvr/auth/deps.py" = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["alembic.*", "jose.*", "passlib.*", "ebooklib.*", "fitz", "fitz.*", "mobi", "mobi.*", "PIL", "PIL.*"]
ignore_missing_imports = true
